refactor(course-page-edit): move block selection to zustand
