+++ src/features/course/features/page-edit/model/editor-store.ts
@@ -1,6 +1,11 @@
 import { createStore, type StoreApi } from 'zustand/vanilla';
 
 import {
+  createCoursePageBlockSelectionItem,
+  type CoursePageBlockSelection,
+  type CoursePageBlockSelectionTarget,
+} from '@/features/course/features/page-edit/model/block-selection';
+import {
   areCoursePageBlockInsertPanelStatesEqual,
   getCoursePageBlockInsertPanelTargetKey,
   getNextCoursePageBlockInsertPanelState,
@@ -14,6 +19,7 @@ export type {
 } from '@/features/course/features/page-edit/model/insert-panel-targeting';
 
 export type CoursePageEditStoreState = {
+  blockSelection: CoursePageBlockSelection;
   contentEditorContainer: HTMLElement | null;
   hoveredInsertPanelTargetKey: string | null;
   insertPanelState: CoursePageBlockInsertPanelState;
@@ -21,8 +27,11 @@ export type CoursePageEditStoreState = {
 };
 
 export type CoursePageEditStoreActions = {
+  clearBlockSelection: () => void;
   hideInsertPanel: () => void;
   hideInsertPanelPreview: () => void;
+  replaceBlockSelection: (targets: CoursePageBlockSelectionTarget[]) => void;
+  selectOnlyBlock: (target: CoursePageBlockSelectionTarget) => void;
   setContentEditorContainer: (container: HTMLElement | null) => void;
   setInsertPanelHovered: (isHovered: boolean) => void;
   showInsertPanelPreview: (cursorY: number) => void;
@@ -105,10 +114,17 @@ export function createCoursePageEditStore(): CoursePageEditStoreApi {
     }
 
     return {
+      blockSelection: { status: 'none' },
       contentEditorContainer: null,
       hoveredInsertPanelTargetKey: null,
       insertPanelState: getHiddenInsertPanelState(),
       isInsertPanelHovered: false,
+      clearBlockSelection: () =>
+        set((state) =>
+          state.blockSelection.status === 'none'
+            ? state
+            : { blockSelection: { status: 'none' } }
+        ),
       hideInsertPanel: () => {
         cancelPendingPreview();
         setInsertPanelState(getHiddenInsertPanelState());
@@ -117,6 +133,33 @@ export function createCoursePageEditStore(): CoursePageEditStoreApi {
         if (!get().isInsertPanelHovered) {
           get().hideInsertPanel();
         }
+      },
+      replaceBlockSelection: (targets) => {
+        const items = targets.map(createCoursePageBlockSelectionItem);
+
+        if (items.length === 0) {
+          get().clearBlockSelection();
+          return;
+        }
+
+        set({
+          blockSelection: {
+            status: 'selected',
+            anchorKey: items.at(-1)?.key ?? items[0].key,
+            items,
+          },
+        });
+      },
+      selectOnlyBlock: (target) => {
+        const item = createCoursePageBlockSelectionItem(target);
+
+        set({
+          blockSelection: {
+            status: 'selected',
+            anchorKey: item.key,
+            items: [item],
+          },
+        });
       },
       setContentEditorContainer: (container) =>
         set((state) =>
